requirements.txt
torch>=2.0.0
torchvision
transformers>=4.30.0
datasets
pillow
easyocr
pytesseract
pandas
numpy
tqdm
scikit-learn
sentencepiece
seqeval
pyyaml
python-multipart
uvicorn
fastapi
pydantic
pdf2image
python-dotenv
python-multipart
